+++ src/parser/mrt/mrt_elem.rs
@@ -110,7 +110,8 @@ fn get_relevant_attributes(
             AttributeValue::OriginatorId(_)
             | AttributeValue::Clusters(_)
             | AttributeValue::Development(_)
-            | AttributeValue::LinkState(_) => {}
+            | AttributeValue::LinkState(_)
+            | AttributeValue::TunnelEncapsulation(_) => {}
         };
     }